requirements-lin.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2024.2
arrow==1.3.0
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
django-cors-headers==4.3.1
django-debug-toolbar==4.4.3
django-environ==0.11.2
django-import-export==4.1.1
django-redis==5.4.0
django-rest-authtoken==2.1.4
django-rq==2.10.2
python-dateutil==2.9.0.post0
redis==5.0.8
rq==1.16.2
six==1.16.0
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240316
